# api/src/routes/v1/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gomate API",
    "description": "Gomate 公开 API — 支持 session cookie 和 API Key 两种认证方式。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.gomate.live",
      "description": "生产环境"
    },
    {
      "url": "https://api-staging.gomate.live",
      "description": "Staging 环境"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API Key（格式：gm_live_<key>），通过 P1b UI 或 POST /auth/api-key/create 创建"
      },
      "SessionCookie": {
        "type": "http",
        "scheme": "bearer",
        "description": "Web 端 session cookie（登录后自动携带）"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "sessionCookie": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
...
      "LocationDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string",
            "nullable": true
          },
          "coverImage": {
            "type": "string",
            "nullable": true
          },
          "cityId": {
            "type": "string",
            "nullable": true
          },
          "difficulty": {
            "type": "string",
            "nullable": true
          },
          "durationMin": {
            "type": "integer",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "bestSeason": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "coordinates": {
            "$ref": "#/components/schemas/Coordinates"
          },
          "gearEssential": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "gearOptional": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            }
          }
        }
      },
      "Coordinates": {
        "type": "object",
        "properties": {
          "lat": {
            "type": "number"
          },
          "lng": {
            "type": "number"
          }
        }
      },
      "Tag": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "StoryCard": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "coverImage": {
            "type": "string",
            "nullable": true
          },
          "summary": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "viewCount": {
            "type": "integer"
          },
          "likeCount": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "author": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "image": {
                "type": "string",
                "nullable": true
              }
            },
            "nullable": true
          },
          "location": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "slug": {
                "type": "string",
                "nullable": true
              }
            },
            "nullable": true
          }
        }
      },
      "StoryDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/StoryCard"
          },
          {
            "type": "object",
            "properties": {
              "isLiked": {
                "type": "boolean"
              },
              "tags": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Tag"
                }
              }
            }
          }
        ]
      }
    }
  },
  "paths": {
...
    "/v1/locations": {
      "get": {
        "operationId": "listLocations",
        "summary": "获取地点列表",
        "description": "公开读端点，支持分页、城市和关键词过滤。",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 12,
              "maximum": 100
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "地点名称搜索"
          },
          {
            "name": "cityId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "按城市 ID 过滤"
          }
        ],
        "responses": {
          "200": {
            "description": "地点列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "locations": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LocationDetail"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createLocation",
        "summary": "创建地点",
        "description": "认证端点（session 或 x-api-key），需要 Idempotency-Key header。",
        "security": [
          {
            "apiKeyAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "地点名称"
                  },
                  "slug": {
                    "type": "string",
                    "description": "slug（可选）"
                  },
                  "type": {
                    "type": "string"
                  },
                  "subtitle": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string",
                    "description": "描述"
                  },
                  "address": {
                    "type": "string"
                  },
                  "cityId": {
                    "type": "string",
                    "description": "城市ID"
                  },
                  "cityName": {
                    "type": "string"
                  },
                  "coverImage": {
                    "type": "string",
                    "format": "uri",
                    "description": "封面图 URL"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uri"
                    }
                  },
                  "coordinates": {
                    "type": "object",
                    "properties": {
                      "lat": {
                        "type": "number"
                      },
                      "lng": {
                        "type": "number"
                      }
                    }
                  },
                  "bestSeason": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "gearEssential": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "gearOptional": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": ["name", "description", "cityId", "coverImage"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "地点创建成功"
          },
          "400": {
            "description": "缺少 Idempotency-Key 或输入无效"
          },
          "401": {
            "description": "未认证"
          },
          "404": {
            "description": "城市不存在"
          }
        }
      }
    },
    "/v1/locations/{id}": {
      "get": {
        "operationId": "getLocationDetail",
        "summary": "获取地点详情",
        "description": "公开读端点，支持按 ID 或 slug 查询。返回坐标、标签、装备建议等完整信息。",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "地点 ID 或 slug"
          }
        ],
        "responses": {
          "200": {
            "description": "地点详情",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "location": {
                      "$ref": "#/components/schemas/LocationDetail"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "地点不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/stories": {
      "get": {
        "operationId": "listStories",
        "summary": "获取已发布故事列表",
        "description": "公开读端点，仅返回 published 状态的故事。",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10,
              "maximum": 20
            }
          },
          {
            "name": "locationId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "按地点 ID 过滤"
          }
        ],
        "responses": {
          "200": {
            "description": "故事列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "stories": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/StoryCard"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createStory",
        "summary": "发布故事",
        "description": "认证端点（session 或 x-api-key），需要 Idempotency-Key header。",
        "security": [
          {
            "apiKeyAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "标题",
                    "maxLength": 100
                  },
                  "summary": {
                    "type": "string",
                    "description": "摘要",
                    "maxLength": 150
                  },
                  "content": {
                    "type": "string",
                    "description": "内容",
                    "maxLength": 10000
                  },
                  "coverImage": {
                    "type": "string",
                    "format": "uri"
                  },
                  "locationId": {
                    "type": "string",
                    "description": "关联地点ID"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "maxItems": 10
                  }
                },
                "required": ["title", "summary", "content", "locationId"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "故事发布成功"
          },
          "400": {
            "description": "缺少 Idempotency-Key 或输入无效"
          },
          "401": {
            "description": "未认证"
          },
          "404": {
            "description": "地点不存在"
          }
        }
      }
    },
    "/v1/stories/{id}": {
      "get": {
        "operationId": "getStoryDetail",
        "summary": "获取故事详情",
        "description": "公开读端点。认证用户可看到 isLiked 状态；draft/hidden 仅对作者本人或 admin 可见。",
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "故事 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "故事详情",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "story": {
                      "$ref": "#/components/schemas/StoryDetail"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "故事不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}